+++ src/main/scala/totoro/ocelot/brain/nbt/NBTTagInt.java
@@ -50,7 +50,7 @@ public class NBTTagInt extends NBTNumber {
     }
 
     public long getLong() {
-        return (long) this.data;
+        return this.data;
     }
 
     public int getInt() {
@@ -66,7 +66,7 @@ public class NBTTagInt extends NBTNumber {
     }
 
     public double getDouble() {
-        return (double) this.data;
+        return this.data;
     }
 
     public float getFloat() {
