+++ src/main/scala/totoro/ocelot/brain/nbt/NBTTagShort.java
@@ -50,7 +50,7 @@ public class NBTTagShort extends NBTNumber {
     }
 
     public long getLong() {
-        return (long) this.data;
+        return this.data;
     }
 
     public int getInt() {
@@ -66,10 +66,10 @@ public class NBTTagShort extends NBTNumber {
     }
 
     public double getDouble() {
-        return (double) this.data;
+        return this.data;
     }
 
     public float getFloat() {
-        return (float) this.data;
+        return this.data;
     }
 }
